поправил dto

+++ src/test/java/faang/school/analytics/listener/donation_analysis/FundRaisedEventListenerTest.java
@@ -14,6 +14,7 @@ import org.mockito.MockitoAnnotations;
 import org.springframework.data.redis.connection.Message;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import static org.mockito.Mockito.*;
@@ -47,7 +48,7 @@ public class FundRaisedEventListenerTest {
         byte[] messageBody = "{\"userId\":1,\"amount\":100,\"timestamp\":\"2024-12-11T10:00:00\"}".getBytes();
         when(message.getBody()).thenReturn(messageBody);
 
-        FundRaisedEvent event = new FundRaisedEvent(1L, 1L, EventType.POST_VIEW, LocalDateTime.now());
+        FundRaisedEvent event = new FundRaisedEvent(1L, BigDecimal.valueOf(1), EventType.POST_VIEW, LocalDateTime.now());
         when(objectMapper.readValue(messageBody, FundRaisedEvent.class)).thenReturn(event);
 
         AnalyticsEvent analyticsEvent = new AnalyticsEvent();
